resume-screening-bot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.6.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==0.8.11
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
transformers==4.35.2
torch==2.9.1
sentence-transformers==2.2.2
huggingface-hub==0.16.4
SpeechRecognition==3.10.0
numpy==1.26.4
pyttsx3==2.90
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
aiofiles==23.2.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
fastapi-cors==0.0.6
